requirements.txt
flask
flask-cors
langchain
langchain-community
langchain-huggingface
sentence-transformers
pinecone-client
openai
psycopg2-binary
python-dotenv
gunicorn